organize/optimize e2e tests
