+++ end-to-end-tests/package.json
@@ -4,15 +4,15 @@
   "description": "Dependencies for running tests with webdriverio and browserstack",
   "main": "index.js",
   "scripts": {
-    "test-travis": "rm -rf screenshots/diff/; wdio wdio/travis.conf.js || bash image-compare/upload_failed_screenshots.sh",
     "test-webdriver-manager": "rm -rf screenshots/diff/; cd .. && ./scripts/env_vars.sh && eval \"$(./scripts/env_vars.sh)\" && cd end-to-end-tests && (wdio wdio/webdriver-manager.conf.js || bash image-compare/upload_failed_screenshots.sh)",
-    "e2e-local": "rm -rf screenshots/screen/; rm -rf screenshots/reference/; rm -rf screenshots/diff/; cd .. && ./scripts/env_vars.sh && eval \"$(./scripts/env_vars.sh)\" && cd end-to-end-tests && (wdio wdio/webdriver-manager.conf.js || bash image-compare/upload_failed_screenshots.sh)"
-
+    "e2e": "yarn run test-webdriver-manager",
+    "clear": "rm -rf screenshots/screen/; rm -rf screenshots/reference/; rm -rf errorShots/; rm -rf screenshots/diff/;"
   },
   "author": "",
   "license": "",
   "dependencies": {
     "browserstack-local": "1.3.0",
+    "clipboardy": "^2.0.0",
     "wdio-browserstack-service": "0.1.4",
     "wdio-junit-reporter": "^0.3.1",
     "wdio-mocha-framework": "0.5.10",
