+++ end-to-end-tests/specs/mutationMapperTool.spec.js
@@ -173,7 +173,7 @@ describe('Mutation Mapper Tool', function() {
 
             // mutations table should be visiable after oncokb icon shows up,
             // also need to wait for mutations to be sorted properly
-            browser.waitForVisible("tr:nth-child(1) [data-test=oncogenic-icon-image]",20000);
+            browser.waitForVisible("tr:nth-child(1) [data-test=oncogenic-icon-image]",30000);
 
             const mutationsT790M = browser.getText('.//*[text()[contains(.,"T790M")]]')
             assert.equal(mutationsT790M.length, 2, "there should be two samples with a T790M mutation");
