+++ end-to-end-tests/specs/specUtils.js
@@ -1,3 +1,5 @@
+const clipboardy = require('clipboardy');
+
 function waitForOncoprint(timeout) {
     browser.pause(100); // give oncoprint time to disappear
     browser.waitUntil(()=>{
@@ -125,6 +127,14 @@ function setInputText(selector, text){
     browser.setValue(selector, '\uE003'.repeat(browser.getValue(selector).length) + text);
 }
 
+function pasteToElement(elementSelector, text){
+
+    clipboardy.writeSync(text);
+    browser.setValue(elementSelector, ["Shift","Insert"]);
+
+}
+
+
 module.exports = {
     waitForOncoprint: waitForOncoprint,
     goToUrlAndSetLocalStorage: goToUrlAndSetLocalStorage,
@@ -141,4 +151,7 @@ module.exports = {
     setOncoprintMutationsMenuOpen: setOncoprintMutationsMenuOpen,
     getNthOncoprintTrackOptionsElements: getNthOncoprintTrackOptionsElements,
     setInputText: setInputText,
+    pasteToElement: pasteToElement
 };
+
+
